+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+-   ON-3931 # Added attributes example for dynamic form element options
+
 ### 3.2.1 (2019-03-14)
 
 ### Dependencies